Update README.md
I'm the GitHub Learning Lab bot and I'm here to help guide you in your journey to learn and master the various topics covered in this course. I will be using Issue and Pull Request comments to communicate with you. In fact, I already added a pull request for everyone to check.

Co-Authored-By: github-learning-lab[bot] <37936606+github-learning-lab[bot]@users.noreply.github.com>